+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]## [0.1.2.3b3]
+- dash_app.HostEnvironment is now an IntEnum to allow integer compare
 
 ## [0.1.2.3b2] - 2023-03-20
 ### Fixed